requirements.txt
fastapi
uvicorn
python-dotenv
xbox-webapi
requests_oauthlib
itsdangerous